Refactor: modificando estrutura e adicionando languages
