+++ src/utils/i18next/languages/pt/common.json
@@ -149,7 +149,8 @@
       "button": {
         "resend": "Reenviar E-mail",
         "edit": "Editar",
-        "disable": "Desabilitar"
+        "disable": "Desabilitar",
+        "backButton": "Botão voltar"
       }
     },
     "toggleButton": {
